+++ ft_memcmp.c
@@ -6,13 +6,15 @@
 /*   By: pnguyen- <pnguyen-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 17:01:00 by pnguyen-          #+#    #+#             */
-/*   Updated: 2023/11/09 19:13:19 by pnguyen-         ###   ########.fr       */
+/*   Updated: 2023/11/11 16:39:00 by pnguyen-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_memcmp(const void *s1, const void *s2, unsigned int n)
+#include <stddef.h>
+
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	c1;
 	unsigned char	c2;
